pyproject.toml
[tool.poetry]
name = "poirot"
version = "0.1.0"
description = ""
authors = ["zhiyuan <zhiyuan@dao42.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
tqdm = "^4.66.4"
openai = "^1.35.7"
tantivy = "^0.22.0"
diskcache = "^5.6.3"
tiktoken = "^0.7.0"
tree-sitter = "0.21.3"
tree-sitter-languages = "^1.10.2"
tree-sitter-javascript = "^0.21.3"
tree-sitter-python = {git = "https://github.com/tree-sitter/tree-sitter-python"}
backoff = "^2.2.1"
numpy = "^2.0.0"
python-dotenv = "^1.0.1"
tree-sitter-go = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
